backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-elasticsearch-dsl>=8.2",
    "django-import-export>=4.3.7",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "elasticsearch>=8.18.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-debug-toolbar>=5.2.0",
]
prod = [
    "gunicorn>=23.0.0",
]

